List and Chart screens

+++ app/src/main/java/br/com/lno/skillmeter/model/Skill.kt
@@ -17,6 +17,6 @@ data class Skill(
     val name: String,
 
     @ColumnInfo(name = "level")
-    val level: Int
+    val level: Float
 
 ) : Serializable
